requirements.txt
asgiref==3.8.1
babel==2.17.0
Django==5.1.9
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-tenants==3.7.0
djangorestframework==3.16.0
docxcompose==1.4.0
docxtpl==0.20.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
pillow==11.2.1
psycopg2==2.9.10
python-docx==1.1.2
python-dotenv==1.1.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2